// src/lib/api/generated/models/subscriptionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

/**
 * The status of the subscription
 */
export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


export const SubscriptionStatus = {
  active: 'active',
  paused: 'paused',
  canceled: 'canceled',
} as const;