// src/lib/api/generated/models/orderPaymentMethod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Marketplace API
 * API for Village Website & Marketplace
 * OpenAPI spec version: 1.0.0
 */

export type OrderPaymentMethod = typeof OrderPaymentMethod[keyof typeof OrderPaymentMethod];


export const OrderPaymentMethod = {
  card: 'card',
  snap: 'snap',
} as const;